feat: 멤버데이터 추가
